Update README.md
I try to implement a simple model based on the Kaggle Dataset. 

The objective of this model is   to identify stocks  who are likely to change daily .The model uses historical data on stocks and tries to find some similarity with existing stock trades  If some similarity is found those stocks  are classified as potential stocks.  The project used visualization using different python packages to analyze the dataset and also uses ML to predict the future situation.ML has a good ability to understand human activities and behave like them. On starting the project it introduces with the table content and imported packages. 

Stocks prediction model in stock market Recommendation on how to set up a  model for a finance company.